+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fixed bad sort step from initialization
 - Fixed columnPicker to prevent the mutation of the props
+- Fixed the no uniqueness of widgets ids by replacing them by class
 
 ## [0.15.0] - 2020-03-24